azure_function/requirements.txt
azure-functions==1.17.0
requests==2.31.0
